# Conflicts:
#	MOD4/BE/dist/middleware/checkRole.js.map
#	MOD4/BE/dist/src/router/blog-router.js.map
#	MOD4/BE/dist/tsconfig.build.tsbuildinfo
#	MOD4/FE/src/home.js

+++ MOD4/FE/src/home.js
@@ -132,77 +132,74 @@ function showNav() {
     <button onclick="showFormAdd()">Add</button>
     <button onclick="showHome()">Home</button>
     <button onclick="logout()">logout</button>
-   
     <input type="search" id="search" placeholder="Enter name" onkeyup="searchProduct(this.value)">
     `)} else {
             $('#nav').html(`
-    <button onclick="showHome()">Home</button>
+
+    <button onclick="userManager()">User Manager</button>
     <button onclick="logout()">logout</button>
+    <button onclick="showHome()">Home</button>
     <input type="search" id="search" placeholder="Enter name" onkeyup="searchProduct(this.value)">
     `)
         }
 
     } else {
         $('#nav').html(`
-    <svg version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
-\t width="800px" height="600px" viewBox="0 0 800 600" enable-background="new 0 0 800 600" xml:space="preserve">
-<linearGradient id="SVGID_1_" gradientUnits="userSpaceOnUse" x1="174.7899" y1="186.34" x2="330.1259" y2="186.34" gradientTransform="matrix(0.8538 0.5206 -0.5206 0.8538 147.9521 -79.1468)">
-\t<stop  offset="0" style="stop-color:#FFC035"/>
-\t<stop  offset="0.221" style="stop-color:#F9A639"/>
-\t<stop  offset="1" style="stop-color:#E64F48"/>
-</linearGradient>
-<circle fill="url(#SVGID_1_)" cx="266.498" cy="211.378" r="77.668"/>
-<linearGradient id="SVGID_2_" gradientUnits="userSpaceOnUse" x1="290.551" y1="282.9592" x2="485.449" y2="282.9592">
-\t<stop  offset="0" style="stop-color:#FFA33A"/>
-\t<stop  offset="0.0992" style="stop-color:#E4A544"/>
-\t<stop  offset="0.9624" style="stop-color:#00B59C"/>
-</linearGradient>
-<circle fill="url(#SVGID_2_)" cx="388" cy="282.959" r="97.449"/>
-<linearGradient id="SVGID_3_" gradientUnits="userSpaceOnUse" x1="180.3469" y1="362.2723" x2="249.7487" y2="362.2723">
-\t<stop  offset="0" style="stop-color:#12B3D6"/>
-\t<stop  offset="1" style="stop-color:#7853A8"/>
-</linearGradient>
-<circle fill="url(#SVGID_3_)" cx="215.048" cy="362.272" r="34.701"/>
-<linearGradient id="SVGID_4_" gradientUnits="userSpaceOnUse" x1="367.3469" y1="375.3673" x2="596.9388" y2="375.3673">
-\t<stop  offset="0" style="stop-color:#12B3D6"/>
-\t<stop  offset="1" style="stop-color:#7853A8"/>
-</linearGradient>
-<circle fill="url(#SVGID_4_)" cx="482.143" cy="375.367" r="114.796"/>
-<linearGradient id="SVGID_5_" gradientUnits="userSpaceOnUse" x1="365.4405" y1="172.8044" x2="492.4478" y2="172.8044" gradientTransform="matrix(0.8954 0.4453 -0.4453 0.8954 127.9825 -160.7537)">
-\t<stop  offset="0" style="stop-color:#FFA33A"/>
-\t<stop  offset="1" style="stop-color:#DF3D8E"/>
-</linearGradient>
-<circle fill="url(#SVGID_5_)" cx="435.095" cy="184.986" r="63.504"/>
-</svg>
-
-
-<div class="container">
-  <h2>login</h2>
-  <form>
-    <input type="text" class="email" placeholder="email">
-    <br/>
-    <input type="text" class="pwd" placeholder="password">
-  </form>
-  <a href="#" class="link">
-    forgot your password ?
-  </a>
-  <br/>
-  <button class="register" onclick="showFormRegister()">
-    <span>register</span>
-  </button>
-  <button class="signin" onclick="showFormLogin()">
-    <span>sign in</span>
-  </button>
-  <h3>your registration is complete !    </h3>
-  <h3>your sign in is complete !</h3>
-  <div class="reg"></div>
-  <div class="sig"></div>
-
- 
- 
-</div>
+    <button onclick="showFormLogin()">Login</button>
+    <button onclick="showFormRegister()">Register</button>
     `)
     }
+}
+
+function userManager(){
+    let token = localStorage.getItem('token')
+    if(token){
+        token = JSON.parse(token)
+        // console.log(token.role)
+        $.ajax({
+            type: 'GET',
+            url: 'http://localhost:3000/auth',
+            headers: {
+                'Content-Type': 'application/json',
+                Authorization: 'Bearer ' + token.token
+            },
+            success: (users) => {
+                // console.log(users);
+                let html = '';
+                    users.map(item => {
+                        html += `<tr>
+             <td>${item.id}</td>
+            <td>${item.username}</td>
+            <td>${item.role}</td>
+            <td><button onclick="lock(${item.id})">${item.status}</button></td>
+            <td><button onclick="deleteRemove(${item.id})">Delete</button></td>
+        </tr>`
+                    })
+                    $('#tbody').html(html)
+            }
+        })
+    }
+}
+
+function lock(id) {
+    if(confirm('lock ?')){
+        let token = localStorage.getItem('token')
+        if (token) {
+            token = JSON.parse(token)
+            $.ajax({
+                type: 'PUT',
+                url: `http://localhost:3000/auth/lock/${id}`,
+                headers: {
+                    'Content-Type': 'application/json',
+                    Authorization: 'Bearer ' + token.token
+                },
+                success: () => {
+                    userManager();
+                }
+            })
+        }
+    }
+
 }
 
 function add() {
@@ -238,8 +235,6 @@ function add() {
                     idBlog : idBlog,
                     idCategory : category
                 }
-                // console.log(blogCategory,11111)
-
                 $.ajax({
                     type: 'POST',
                     url: 'http://localhost:3000/blogs/blogCategory',
@@ -249,8 +244,8 @@ function add() {
                     },
                     data: JSON.stringify(blogCategory),
 
-                    success: () => {
-                        // console.log(blogCategory,222222)
+                    success: (blogCategory) => {
+                        console.log(blogCategory,197)
                         showHome()
                     }
                 })
@@ -344,7 +339,7 @@ function uploadImage(e) {
     let uploadTask = storageRef.put(file);
     uploadTask.on(firebase.storage.TaskEvent.STATE_CHANGED,
         function (snapshot) {
-
+            uploader.value = (snapshot.bytesTransferred / snapshot.totalBytes) * 100;
             switch (snapshot.state) {
                 case firebase.storage.TaskState.PAUSED:
                     break;
@@ -423,63 +418,9 @@ function searchProduct(value) {
 
 function showFormLogin() {
     $('#body').html(` 
-             <svg version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
-\t width="800px" height="600px" viewBox="0 0 800 600" enable-background="new 0 0 800 600" xml:space="preserve">
-<linearGradient id="SVGID_1_" gradientUnits="userSpaceOnUse" x1="174.7899" y1="186.34" x2="330.1259" y2="186.34" gradientTransform="matrix(0.8538 0.5206 -0.5206 0.8538 147.9521 -79.1468)">
-\t<stop  offset="0" style="stop-color:#FFC035"/>
-\t<stop  offset="0.221" style="stop-color:#F9A639"/>
-\t<stop  offset="1" style="stop-color:#E64F48"/>
-</linearGradient>
-<circle fill="url(#SVGID_1_)" cx="266.498" cy="211.378" r="77.668"/>
-<linearGradient id="SVGID_2_" gradientUnits="userSpaceOnUse" x1="290.551" y1="282.9592" x2="485.449" y2="282.9592">
-\t<stop  offset="0" style="stop-color:#FFA33A"/>
-\t<stop  offset="0.0992" style="stop-color:#E4A544"/>
-\t<stop  offset="0.9624" style="stop-color:#00B59C"/>
-</linearGradient>
-<circle fill="url(#SVGID_2_)" cx="388" cy="282.959" r="97.449"/>
-<linearGradient id="SVGID_3_" gradientUnits="userSpaceOnUse" x1="180.3469" y1="362.2723" x2="249.7487" y2="362.2723">
-\t<stop  offset="0" style="stop-color:#12B3D6"/>
-\t<stop  offset="1" style="stop-color:#7853A8"/>
-</linearGradient>
-<circle fill="url(#SVGID_3_)" cx="215.048" cy="362.272" r="34.701"/>
-<linearGradient id="SVGID_4_" gradientUnits="userSpaceOnUse" x1="367.3469" y1="375.3673" x2="596.9388" y2="375.3673">
-\t<stop  offset="0" style="stop-color:#12B3D6"/>
-\t<stop  offset="1" style="stop-color:#7853A8"/>
-</linearGradient>
-<circle fill="url(#SVGID_4_)" cx="482.143" cy="375.367" r="114.796"/>
-<linearGradient id="SVGID_5_" gradientUnits="userSpaceOnUse" x1="365.4405" y1="172.8044" x2="492.4478" y2="172.8044" gradientTransform="matrix(0.8954 0.4453 -0.4453 0.8954 127.9825 -160.7537)">
-\t<stop  offset="0" style="stop-color:#FFA33A"/>
-\t<stop  offset="1" style="stop-color:#DF3D8E"/>
-</linearGradient>
-<circle fill="url(#SVGID_5_)" cx="435.095" cy="184.986" r="63.504"/>
-</svg>
-
-
-<div class="container">
-  <h2>login</h2>
-  <form>
-    <input type="text" class="email" placeholder="email">
-    <br/>
-    <input type="text" class="pwd" placeholder="password">
-  </form>
-  <a href="#" class="link">
-    forgot your password ?
-  </a>
-  <br/>
-  <button class="register" onclick="signup()">
-    <span>register</span>
-  </button>
-  <button class="signin" onclick="login()">
-    <span>sign in</span>
-  </button>
-  <h3>your registration is complete !    </h3>
-  <h3>your sign in is complete !</h3>
-  <div class="reg"></div>
-  <div class="sig"></div>
-
- 
- 
-</div>
+             <input type="text" id = "username" placeholder="username"> 
+             <input type="password" id = "password" placeholder="password"> 
+             <button onclick="login()">Login</button>
 `)
 }
 
@@ -500,10 +441,16 @@ function login() {
         data: JSON.stringify(user),
 
         success: (token) => {
-            localStorage.setItem('token', JSON.stringify(token));
+            if(token === "Username is not existed" ||token === 'Password is wrong' ){
+                alert('Can not')
+                showNav();
+            }else{
+                localStorage.setItem('token', JSON.stringify(token));
+                showNav();
+                showHome();
+            }
 
-            showNav();
-            showHome();
+
         }
     })
 }
@@ -516,63 +463,9 @@ function logout() {
 
 function showFormRegister() {
     $('#body').html(` 
-             <svg version="1.1" id="Layer_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
-\t width="800px" height="600px" viewBox="0 0 800 600" enable-background="new 0 0 800 600" xml:space="preserve">
-<linearGradient id="SVGID_1_" gradientUnits="userSpaceOnUse" x1="174.7899" y1="186.34" x2="330.1259" y2="186.34" gradientTransform="matrix(0.8538 0.5206 -0.5206 0.8538 147.9521 -79.1468)">
-\t<stop  offset="0" style="stop-color:#FFC035"/>
-\t<stop  offset="0.221" style="stop-color:#F9A639"/>
-\t<stop  offset="1" style="stop-color:#E64F48"/>
-</linearGradient>
-<circle fill="url(#SVGID_1_)" cx="266.498" cy="211.378" r="77.668"/>
-<linearGradient id="SVGID_2_" gradientUnits="userSpaceOnUse" x1="290.551" y1="282.9592" x2="485.449" y2="282.9592">
-\t<stop  offset="0" style="stop-color:#FFA33A"/>
-\t<stop  offset="0.0992" style="stop-color:#E4A544"/>
-\t<stop  offset="0.9624" style="stop-color:#00B59C"/>
-</linearGradient>
-<circle fill="url(#SVGID_2_)" cx="388" cy="282.959" r="97.449"/>
-<linearGradient id="SVGID_3_" gradientUnits="userSpaceOnUse" x1="180.3469" y1="362.2723" x2="249.7487" y2="362.2723">
-\t<stop  offset="0" style="stop-color:#12B3D6"/>
-\t<stop  offset="1" style="stop-color:#7853A8"/>
-</linearGradient>
-<circle fill="url(#SVGID_3_)" cx="215.048" cy="362.272" r="34.701"/>
-<linearGradient id="SVGID_4_" gradientUnits="userSpaceOnUse" x1="367.3469" y1="375.3673" x2="596.9388" y2="375.3673">
-\t<stop  offset="0" style="stop-color:#12B3D6"/>
-\t<stop  offset="1" style="stop-color:#7853A8"/>
-</linearGradient>
-<circle fill="url(#SVGID_4_)" cx="482.143" cy="375.367" r="114.796"/>
-<linearGradient id="SVGID_5_" gradientUnits="userSpaceOnUse" x1="365.4405" y1="172.8044" x2="492.4478" y2="172.8044" gradientTransform="matrix(0.8954 0.4453 -0.4453 0.8954 127.9825 -160.7537)">
-\t<stop  offset="0" style="stop-color:#FFA33A"/>
-\t<stop  offset="1" style="stop-color:#DF3D8E"/>
-</linearGradient>
-<circle fill="url(#SVGID_5_)" cx="435.095" cy="184.986" r="63.504"/>
-</svg>
-
-
-<div class="container">
-  <h2>login</h2>
-  <form>
-    <input type="text" class="email" placeholder="email">
-    <br/>
-    <input type="text" class="pwd" placeholder="password">
-  </form>
-  <a href="#" class="link">
-    forgot your password ?
-  </a>
-  <br/>
-  <button class="register" onclick="signup()">
-    <span>register</span>
-  </button>
-  <button class="signin" onclick="login()">
-    <span>sign in</span>
-  </button>
-  <h3>your registration is complete !    </h3>
-  <h3>your sign in is complete !</h3>
-  <div class="reg"></div>
-  <div class="sig"></div>
-
- 
- 
-</div>
+             <input type="text" id = "username" placeholder="username"> 
+             <input type="password" id = "password" placeholder="password"> 
+             <button onclick="signup()">Signup</button>
 `)
 }
